requirements.txt
fastapi
uvicorn
tensorflow
transformers
python-multipart
Pillow
numpy